# Faster cell update